+++ obs-pelotense-controle/build-32.1.2/.qt/QtDeployTargets-Debug.cmake
@@ -1,0 +1,3 @@
+set(__QT_DEPLOY_TARGET_obs-pelotense-controle_FILE C:/Users/tsicr/OneDrive/Área de Trabalho/Pelotense-Esportes/obs-pelotense-controle/build-32.1.2/obs-plugins/64bit/Debug/obs-pelotense-controle.dll)
+set(__QT_DEPLOY_TARGET_obs-pelotense-controle_TYPE MODULE_LIBRARY)
+set(__QT_DEPLOY_TARGET_obs-pelotense-controle_RUNTIME_DLLS C:/Users/tsicr/AppData/Local/Temp/opencode/obs-deps-qt6-2025-08-23-x64/bin/Qt6Widgetsd.dll;C:/Users/tsicr/AppData/Local/Temp/opencode/obs-deps-qt6-2025-08-23-x64/bin/Qt6Guid.dll;C:/Users/tsicr/AppData/Local/Temp/opencode/obs-deps-qt6-2025-08-23-x64/bin/Qt6Cored.dll)