+++ obs-pelotese-assets-controle/build-32.1.2/.qt/QtDeployTargets-RelWithDebInfo.cmake
@@ -1,0 +1,3 @@
+set(__QT_DEPLOY_TARGET_obs-pelotese-assets-controle_FILE C:/Users/tsicr/OneDrive/Área de Trabalho/Pelotense-Esportes/obs-pelotese-assets-controle/build-32.1.2/obs-plugins/64bit/RelWithDebInfo/obs-pelotese-assets-controle.dll)
+set(__QT_DEPLOY_TARGET_obs-pelotese-assets-controle_TYPE MODULE_LIBRARY)
+set(__QT_DEPLOY_TARGET_obs-pelotese-assets-controle_RUNTIME_DLLS C:/Users/tsicr/AppData/Local/Temp/opencode/obs-deps-qt6-2025-08-23-x64/bin/Qt6Widgets.dll;C:/Users/tsicr/AppData/Local/Temp/opencode/obs-deps-qt6-2025-08-23-x64/bin/Qt6Gui.dll;C:/Users/tsicr/AppData/Local/Temp/opencode/obs-deps-qt6-2025-08-23-x64/bin/Qt6Core.dll)